projects/inoah-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "inoah-core"
version = "1.0.0"
description = "iNoah Core - Shared infrastructure for the iNoah AI ecosystem"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "Pillow>=9.0.0",
    "chromadb>=0.4.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "mypy"]

[tool.setuptools.packages.find]
where = ["src"]
